requirements.txt
streamlit==1.31.0
openai==1.12.0
selenium==4.15.2
beautifulsoup4==4.12.2
pandas==2.1.3
scikit-learn==1.3.2
python-dotenv==1.0.0
apscheduler==3.10.4
